Split up certificate.rs into multiple files and a module
